configure.ac: add option to disable image output
Even when the gd library is present, it may be desirable to not build
'vnstati' as image output is an optional feature.

Also, when testing for the gd library, use gdImagePng() instead of
gdImageLine() since it's possible that the installed gd library doesn't
have PNG support. In such cases, the test in the configure script passed
(because gdImageLine() is always present), but the build failed with:

  src/vnstati.o: In function `writeoutput':
  vnstati.c:(.text+0x3fc): undefined reference to `gdImagePng'
  collect2: error: ld returned 1 exit status