Autotools: fix case of wrong path being displayed
Besides the prefix exist a few more variables which have influrence here.
They were already used (if set), in which case the wrong path was displayed here.

The output will differ now, in that it displays variable names,
because only one layer of the variables are resolved,
and more than one layer is used now.